requirements.txt
Django == 3.0.4
djangorestframework == 3.11.0
markdown == 3.2.1
django-filter == 2.2.0
django-cors-headers == 3.2.1
django-rest-swagger == 2.2.0
xhtml2pdf == 0.2.4
drf-yasg == 1.17.0
reportlab == 3.5.34
pdfkit ==  0.6.1